1) so-layers now run every parameter through either addslashes of intval, to prevent query-insertion and for pgSql 7.3 compatibility 2) corrected the escapeing of " and ' 3) added some inline-docs

+++ infolog/inc/class.soinfolog.inc.php
@@ -13,6 +13,14 @@
 
 	/* $Id$ */
 
+	/*!
+	@class soinfolog
+	@abstract storage object / db-layer for InfoLog
+	@author Ralf Becker
+	@copyright GPL - GNU General Public License
+	@note all values passed to this class are run either through intval or addslashes to prevent query-inserting
+		and for pgSql 7.3 compatibility
+	*/
 	class soinfolog 				// DB-Layer
 	{
 		var $db,$db2;
@@ -20,9 +28,11 @@
 		var $data = array( );
 		var $filters = array( );
 		var $user;
-		var $maybe_slashes = array (
-			'info_des'=>1,'info_subject'=>1,'info_from'=>1,'info_addr'=>1
-		);
+
+		/*!
+		@function soinfolog
+		@abstract constructor
+		*/
 		function soinfolog( $info_id = 0)
 		{
 			$this->db     = $GLOBALS['phpgw']->db;
@@ -33,8 +43,15 @@
 
 			$this->read( $info_id );
 		}
-				
 
+		/*!
+		@function check_access
+		@abstract checks if user has the $required_rights to access $info_id (private access is handled too)
+		@syntax check_access( $info_id,$required_rights )
+		@param $info_id Id of InfoLog entry
+		@param $required_rights PHPGW_ACL_xyz anded together
+		@returns True if access is granted else False
+		*/
 		function check_access( $info_id,$required_rights )
 		{
 			if ($info_id != $this->data['info_id'])      	// already loaded?
@@ -65,10 +82,13 @@
 			return $access_ok;         
 		}
 
-		// sql to be AND into a query to ensure ACL is respected (incl. _PRIVATE)
-		// filter: none|all    - list all entrys user have rights to see
-		//         private|own - list only his personal entrys
-		//							(incl. those he is responsible for !!!)            
+		/*!
+		@function aclFilter
+		@abstract generate sql to be AND'ed into a query to ensure ACL is respected (incl. _PRIVATE)
+		@param $filter: none|all - list all entrys user have rights to see<br>
+			private|own - list only his personal entrys (incl. those he is responsible for !!!) 
+		@returns the necesary sql
+		*/           
 		function aclFilter($filter = 'none')
 		{
 			ereg('.*(own|privat|all|none).*',$filter,$vars);
@@ -94,8 +114,7 @@
 				}
 				if (count($private_user_list))
 				{               
-					$has_private_access = 'info_owner IN ('.
-												 implode(',',$private_user_list).')';
+					$has_private_access = 'info_owner IN ('.implode(',',$private_user_list).')';
 				}            
 			}
 			$filtermethod = " (info_owner=$this->user"; // user has all rights
@@ -122,6 +141,13 @@
 			return $this->acl_filter[$filter] = $filtermethod;  // cache the filter
 		}      
 	
+		/*!
+		@function statusFilter
+		@abstract generate sql to filter based on the status of the log-entry
+		@syntax statusFilter($filter = '')
+		@param $filter done = done or billed, open = not ()done or billed), offer = offer
+		@returns the necesary sql
+		*/
 		function statusFilter($filter = '')
 		{
 			ereg('.*(done|open|offer).*',$filter,$vars);
@@ -136,6 +162,15 @@
 			return '';
 		}
 
+		/*!
+		@function dateFilter
+		@abstract generate sql to filter based on the start- and enddate of the log-entry
+		@syntax dateFilter($filter = '')
+		@param $filter upcoming = startdate is in the future<br>
+			today startdate < tomorrow<br>
+			overdue enddate < tomorrow
+		@returns the necesary sql
+		*/
 		function dateFilter($filter = '')
 		{
 			ereg('.*(upcoming|today|overdue).*',$filter,$vars);
@@ -146,45 +181,56 @@
 
 			switch ($filter)
 			{
-				case 'upcoming':	return " AND info_startdate >= '$tomorrow'";
-				case 'today':		return " AND info_startdate < '$tomorrow'";
-				case 'overdue':	return " AND (info_enddate != 0 AND info_enddate < '$tomorrow')";
+				case 'upcoming':
+					return " AND info_startdate >= '$tomorrow'";
+				case 'today':
+					return " AND info_startdate < '$tomorrow'";
+				case 'overdue':
+					return " AND (info_enddate != 0 AND info_enddate < '$tomorrow')";
 			}
 			return '';
 		}
 
+		/*!
+		@function init
+		@abstract initialise the internal $this->data to be empty
+		@discussion only non-empty values got initialised
+		*/
 		function init()
 		{
-			$this->data = array( 'info_owner' => $this->user,
-										'info_pri'    => 'normal' );
+			$this->data = array( 
+				'info_owner' => $this->user,
+				'info_pri'   => 'normal' 
+			);
 		}      
 		
+		/*!
+		@function db2data
+		@abstract copy data after a query into $data
+		@syntax db2data(&$data)
+		@param $data array to copy the data
+		@description doesnt do much anymore
+		*/
 		function db2data(&$data)
 		{
 			$data = $this->db->Record;
-			reset($this->maybe_slashes);
-			while (list($key) = each($this->maybe_slashes))
-			{
-				$data[$key] = stripslashes($data[$key]);
-			}
-/*
-			$links = $this->links->get_links('infolog',$this->data['info_id']);
-			while (list($nul,$link) = each($links))
-			{
-				if ($link['app'] == 'addressbook')
-					$data['info_addr_id'] = $link['id'];
-				if ($link['app'] == 'projects')
-					$data['info_proj_id'] = $link['id'];
-				if ($link['app'] == 'calendar')
-					$data['info_event_id'] = $link['id'];
-			}
-*/
 		}
 
+		/*!
+		@function read
+		@abstract read InfoLog entry $info_id
+		@syntax read( $info_id )
+		@param $info_id id of log-entry
+		@description some cacheing is done to prevent multiple reads of the same entry
+		@returns the entry as array
+		*/
 		function read($info_id)		// did _not_ ensure ACL
 		{
+			$info_id = intval($info_id);
+
 			if ($info_id <= 0 || $info_id != $this->data['info_id'] && 
-										(!$this->db->query("select * FROM phpgw_infolog where info_id='$info_id'") ||   !$this->db->next_record())) 
+				(!$this->db->query("select * FROM phpgw_infolog where info_id=$info_id",__LINE__,__FILE__) ||
+				 !$this->db->next_record())) 
 			{
 				$this->init( );
 				return False;
@@ -196,11 +242,21 @@
 			return $this->data;
 		}
 		
+		/*!
+		@function delete
+		@abstract delete InfoLog entry $info_id AND the links to it
+		@syntax delete( $info_id )
+		@param $info_id id of log-entry
+		*/
 		function delete($info_id)  // did _not_ ensure ACL
 		{
-			$this->db->query("delete FROM phpgw_infolog where info_id='$info_id' or info_id_parent='"
-				. "$info_id' AND ((info_access='public' and info_owner != '$this->user')"
-				. " or (info_owner='$this->user'))" ,__LINE__,__FILE__);
+			if (($info_id = intval($info_id)) <= 0)
+			{
+				return;
+			}
+			$this->db->query("delete FROM phpgw_infolog where info_id=$info_id or info_id_parent=$info_id" .
+				"AND ((info_access='public' and info_owner != '$this->user') OR (info_owner=$this->user))",
+				__LINE__,__FILE__);
 				
 			$this->links->unlink(0,'infolog',$info_id);
 
@@ -210,22 +266,37 @@
 			}
 		}
 
+		/*!
+		@function change_delete_owner
+		@abstract changes or deletes entries with a spezified owner (for hook_delete_account)
+		@syntax change_delete_owner( $owner,$new_owner=0 )
+		@param $owner old owner
+		@param $new_owner new owner or 0 if entries should be deleted
+		*/
 		function change_delete_owner($owner,$new_owner=0)  // new_owner=0 means delete
 		{
-			if (!$new_owner)
+			$owner = intval($owner);
+			if (!($new_owner = intval($new_owner)))
 			{
-				$sql = "delete FROM phpgw_infolog where info_owner='$owner'";
-				$sql2 = "update phpgw_infolog set info_responsible='0' where info_responsible='$owner'";
+				$sql = "delete FROM phpgw_infolog where info_owner=$owner";
+				$sql2 = "update phpgw_infolog set info_responsible=0 where info_responsible=$owner";
 			}
 			else
 			{
-				$sql = "update phpgw_infolog set info_owner='$new_owner' where info_owner='$owner'";
-				$sql2 = "update phpgw_infolog set info_responsible='$new_owner' where info_responsible='$owner'";
+				$sql = "update phpgw_infolog set info_owner=$new_owner where info_owner=$owner";
+				$sql2 = "update phpgw_infolog set info_responsible=$new_owner where info_responsible=$owner";
 			}
 			$this->db->query($sql,__LINE__,__FILE__);
 			$this->db->query($sql2,__LINE__,__FILE__);
 		}
 
+		/*!
+		@function write
+		@abstract writes the given $values to InfoLog, a new entry gets created if info_id is not set or 0
+		@syntax write( $values )
+		@param $values array with the data of the log-entry
+		@returns nothing direct, but the info_id gets
+		*/
 		function write($values)  // did _not_ ensure ACL
 		{
 			include(PHPGW_SERVER_ROOT.'/infolog/setup/tables_current.inc.php');
@@ -242,18 +313,23 @@
 					}
 					$this->data[$key] = $val;   // update internal data
 
-					if ($this->maybe_slashes[$key])
+					switch($val['type'])	// protection against query-insertion
 					{
-						$val = addslashes($val);
+						case 'int': case 'auto':
+							$val = intval($val);
+							break;
+						default:
+							$val = "'".$this->db->db_addslashes($val)."'";
+							break;
 					}
 					$cols .= ($cols ? ',' : '').$key;
-					$vals .= ($vals ? ',' : '')."'$val'";
-					$query .= ($query ? ',' : '')."$key='$val'";
+					$vals .= ($vals ? ',' : '').$val;
+					$query .= ($query ? ',' : '')."$key=$val";
 				}
 			}
-			if (($this->data['info_id'] = $values['info_id']) > 0)
+			if (($this->data['info_id'] = intval($values['info_id'])) > 0)
 			{
-				$query = "UPDATE phpgw_infolog SET $query where info_id='".$values['info_id']."'";
+				$query = "UPDATE phpgw_infolog SET $query where info_id='".$this->data['info_id']."'";
 				$this->db->query($query,__LINE__,__FILE__);
 			}
 			else
@@ -264,31 +340,44 @@
 			}
 			// echo "<p>soinfolog.write values= "; _debug_array($values);
 			// echo "<p>soinfolog.write this->data= "; _debug_array($this->data);
-			/*
-			if ($this->data['info_addr_id'])
-				$this->links->link('infolog',$this->data['info_id'],'addressbook',$this->data['info_addr_id']);
-			if ($this->data['info_proj_id'])
-				$this->links->link('infolog',$this->data['info_id'],'projects',$this->data['info_proj_id']);
-			if ($this->data['info_event_id'])
-				$this->links->link('infolog',$this->data['info_id'],'calendar',$this->data['info_event_id']); */
+
+			return $this->data['info_id'];
 		}
 
+		/*!
+		@function anzSubs
+		@abstract count the sub-entries of $info_id
+		@syntax anzSubs( $info_id )
+		@param $info_id id of log-entry
+		@returns the number of sub-entries
+		*/
 		function anzSubs( $info_id )
 		{
-			if ($info_id <= 0)
+			if (($info_id = intval($info_id)) <= 0)
 			{
 				return 0;
 			}
-			$this->db->query('select count(*) FROM phpgw_infolog where '.
-								  "info_id_parent=$info_id",__LINE__,__FILE__);
+			$this->db->query("select count(*) FROM phpgw_infolog where info_id_parent=$info_id",__LINE__,__FILE__);
 
 			$this->db->next_record();
 
 			return $this->db->f(0);
 		}
 
-		function search($order,$sort,$filter,$cat_id,$query,$action,$action_id,
-							 $ordermethod,&$start,&$total)
+		/*!
+		@function search
+		@abstract searches InfoLog for a certain pattern in $query
+		@syntax search( $order,$sort,$filter,$cat_id,$query,$action,$action_id,$ordermethod,&$start,&$total )
+		@param $order comma-separated list of columns to order the result (no 'ORDER BY'), eg. 'info_subject DESC'
+		@param $sort comma-separated list of columns to to sort by (incl. 'SORT BY') or ''
+		@param $filter string with combination of acl-, date- and status-filters, eg. 'own-open-today' or ''
+		@param $cat_id category to use or 0
+		@param $query pattern to search, search is done in info_from, info_subject and info_des
+		@param $action / $action_id if only entries linked to a specified app/entry show be used
+		@param &$start, &$total nextmatch-parameters will be used and set if query returns less entries
+		@returns array with id's as key of the matching log-entries
+		*/
+		function search($order,$sort,$filter,$cat_id,$query,$action,$action_id,$ordermethod,&$start,&$total)
 		{
 			//echo "<p>soinfolog.search(action='$action/$action_id')</p>\n";
 			$action2app = array(
@@ -311,7 +400,7 @@
 			}
 			if ($order)
 			{
-			  $ordermethod = 'ORDER BY ' . $order . ' ' . $sort;
+			  $ordermethod = 'ORDER BY ' . $this->db->db_addslashes($order) . ' ' . $this->db->db_addslashes($sort);
 			}
 			else
 			{
@@ -322,12 +411,13 @@
 			$filtermethod .= $this->dateFilter($filter);
 			// echo "<p>filtermethod='$filtermethod'</p>";
 
-			if ($cat_id)
+			if (intval($cat_id))
 			{
-			  $filtermethod .= " AND info_cat='$cat_id' ";
+			  $filtermethod .= ' AND info_cat='.intval($cat_id).' ';
 			}
 			if ($query)			  // we search in _from, _subject and _des for $query
 			{
+				$query = $this->db->db_addslashes($query);
 				$sql_query = "AND (info_from like '%$query%' OR info_subject ".
 								 "LIKE '%$query%' OR info_des LIKE '%$query%') ";
 			}
@@ -340,7 +430,7 @@
 			}
 			$ids = array( );
 			if ($action == '' || $action == 'sp' || count($links))
-         {
+			{
 				$this->db->query($sql="SELECT COUNT(*) FROM phpgw_infolog i WHERE ($filtermethod $pid $sql_query) $link_extra",__LINE__,__FILE__);
 				
 				$this->db->next_record();
